Update README.md
The README now has a description about what this Repository will be used for. It is used for my DIG2500C Fundamentals of Interactive Design class. All of the projects done in this course will be stored here. This is the first true assignment of the class, and one of the most important parts of a project.